MAINT: move code and fix compile errors.
