+++ cpp/examples/Sara/ChessboardDetection/Utilities/ImageOrVideoReader.cpp
