+++ cpp/examples/Sara/ChessboardDetection/Utilities/ImageOrVideoReader.hpp
