+++ cpp/examples/Sara/ChessboardDetection/find_chessboard_corners.cpp
@@ -17,9 +17,8 @@
 
 #include <DO/Sara/Graphics.hpp>
 
+#include <DO/Sara/ChessboardDetection/ChessboardDetector.hpp>
 #include <DO/Sara/ImageProcessing/FastColorConversion.hpp>
-
-#include <drafts/ChessboardDetection/ChessboardDetector.hpp>
 
 #include "Utilities/ImageOrVideoReader.hpp"
