+++ cpp/src/DO/Sara/Calibration/CMakeLists.txt
@@ -5,5 +5,3 @@ add_library(
 target_link_libraries(DO_Sara_Calibration PUBLIC DO::Sara::Core
                                                  DO::Sara::Graphics)
 set_property(TARGET DO_Sara_Calibration PROPERTY FOLDER "Libraries/Sara")
-
-add_subdirectory(tool)
