+++ cpp/src/DO/Sara/Calibration/Chessboard.cpp
@@ -11,7 +11,7 @@
 
 //! @file
 
-#include <drafts/Calibration/Chessboard.hpp>
+#include <DO/Sara/Calibration/Chessboard.hpp>
 
 
 namespace DO::Sara {
