+++ cpp/src/DO/Sara/Calibration/Chessboard.hpp
@@ -13,7 +13,7 @@
 
 #pragma once
 
-#include <drafts/ChessboardDetection/ChessboardDetector.hpp>
+#include <DO/Sara/ChessboardDetection/ChessboardDetector.hpp>
 
 #include <optional>
