+++ cpp/src/DO/Sara/ChessboardDetection/CircularProfileExtractor.hpp
