+++ cpp/src/DO/Sara/ChessboardDetection/EllipseFitting.hpp
