+++ cpp/src/DO/Sara/ChessboardDetection/NonMaximumSuppression.hpp
