+++ cpp/src/DO/Sara/ChessboardDetection/SquareReconstruction.cpp
@@ -9,7 +9,7 @@
 // you can obtain one at http://mozilla.org/MPL/2.0/.
 // ========================================================================== //
 
-#include "SquareReconstruction.hpp"
+#include <DO/Sara/ChessboardDetection/SquareReconstruction.hpp>
 
 
 namespace DO::Sara {
