+++ cpp/src/DO/Sara/ChessboardDetection/notes.md
